+++ Minecraft/src/common/math.h
@@ -6,7 +6,7 @@ namespace Minecraft
 	{
 	public:
 		static float Intbound(float s, float ds);
-		static int Signum(float x);
+		static float Signum(float x);
 		static float Mod(float value, int modulus);
 
 	};
