+++ Minecraft/src/game/chunk.h
@@ -4,7 +4,6 @@
 #include <glm/glm.hpp>
 #include "block.h"
 
-
 #define CHUNK_SIZE 16
 #define TOTAL_VERTICES CHUNK_SIZE * CHUNK_SIZE * CHUNK_SIZE * 36
 
@@ -13,15 +12,15 @@ namespace Minecraft
 	class Chunk
 	{
 	public:
-		Chunk(const glm::vec3& position);
+		Chunk(const glm::ivec3& position);
 		Chunk() = delete;
 		Chunk(Chunk& other) = delete;
 
-		inline void SetPoisition(const glm::vec3& position) { m_Position = position; }
+		inline void SetPoisition(const glm::ivec3& position) { m_Position = position; }
 
-		inline uint16_t GetBlock(const glm::vec3& position)  const
+		inline uint16_t GetBlock(const glm::ivec3& position)  const
 		{
-			return GetBlock((int)position.x, (int)position.y, (int)position.z);
+			return GetBlock(position.x, position.y, position.z);
 		}
 
 		inline uint16_t GetBlock
@@ -29,18 +28,16 @@ namespace Minecraft
 		{
 			return m_Blocks[x][y][z];
 		}
-		inline void SetBlock(const glm::vec3& position, uint8_t type)
+		inline void SetBlock(const glm::ivec3& position, uint8_t type)
 		{
-			SetBlock((int)position.x, (int)position.y, (int)position.z, type);
+			SetBlock(position.x, position.y, position.z, type);
 		}
 		void SetBlock(int x, int y, int z, uint8_t type)
 		{
 			m_Blocks[x][y][z] = type;
-			//if(type == 0)
-			//MC_INFO("Wut");
 			m_Changed = true;
 		}
-		inline glm::vec3 GetPosition() const { return m_Position; }
+		inline glm::ivec3 GetPosition() const { return m_Position; }
 
 		void Update(vertex* res);
 		inline glm::mat4& GetTransformationMatrix()
@@ -60,7 +57,7 @@ namespace Minecraft
 		uint32_t m_Elements = 0;
 		changed m_Changed = true;
 
-		glm::vec3 m_Position;
+		glm::ivec3 m_Position;
 		uint16_t m_Blocks[CHUNK_SIZE][CHUNK_SIZE][CHUNK_SIZE];
 	};
 }
